refacto(assigned_user): add func to display and edit assigned users
This commit adds the ability to display and edit assigned users for tasks. The `ShowTabController` now has a method to retrieve all users, which is used by both the `ProjectTabService` and `PopupService` classes to display a list of assignable users. The `TaskDetailsController` now has a button that opens a popup allowing the user to change the assigned user for a task. Additionally, the `Task` class has been updated with a new method that retrieves an assigned user object instead of just returning their name as a string. Finally, the `UserService` class has been updated with a new method that creates an HBox containing a ComboBox of all assignable users and a delete button for removing an assigned user from a task.