fix(task): Update TaskApiClient to allow updating assigned user
This commit updates the `TaskApiClient` class to allow for updating the assigned user of a task. The `updateTask` method now takes an additional parameter, `userId`, which specifies the ID of the new assigned user. If this parameter is not provided, it defaults to the current assigned user's ID. Additionally, in `ShowTabController`, we update tasks with their new assigned users when they are moved between columns. Finally, in `TaskDetailsController`, we use `TaskApiClient.getInstance().getTaskById()` to retrieve a fresh copy of a task before showing its details and allowing changes to be made.