feat: Add delete and leave project functionality
This commit adds the ability to delete and leave projects. The `deleteProject` method in `ProjectApiClient` now takes a `Project` object as an argument instead of an ID, and sends a DELETE request to the server using the `HttpClientApi`. Similarly, a new method called `leaveProject` has been added to the interface and implementation classes. Additionally, this commit includes a minor change in `CreateTabController`, where the `resetFields` method is called after joining a table. Finally, there is also some refactoring done in the code for better readability.