fix: Add error message when no tables are found
This commit adds an error message to be displayed when there are no tables found. The error message is displayed in red and has a smaller font size than the other labels. The changes were made to the ShowAllTabController, EString enum, styles.css file, and show_all_tab.fxml file.